fence_vmware: Fix process of special characters (like ')
This is only one agent using pexpect.run function (powerful
replacement for popen function). It is able to parse strings
in very shell way (so string closed in ' or " is  processed
with spaces). Problem is, if user has password/user name/...
with this characters.

Patch fixing bug by adding quote_for_run function, which just
replace every ' by '\''. So to pexpect.run is passed right
string.