fence_agents: #487501 - Exceptions in fencing agents
Several of new fencing agent does not work when bad options are entered (bad
password, non-existent outlet number, ...). This is not a problem. Unfortunately
python exceptions are thrown and we don't care about catching
them. Such output is very usefull for debugging (as people tends to send it)
but it will be better if the user will see that the error is in the input data.

Only bigger change is in APC. In other cases only catching exception was added.