fencing: New option --retry-on <N>
New firmware (v1.70+) for ilo2 is much slower then before. Option --retry-on=X (on stdin retry_on=X)
is introduced and it attempts to send power on command to machine and wait for results X times. Default for
fence agents is 1 so they won't be affected; fence_ilo default value is 3.

Resolves: #bz507514 (timeout options are needed before applying this patch separately)