fence_scsi: fix regular expression for grep
When fence_scsi checks to see if a specific keys is registered for a
specific device, it uses grep to look through a list of all registered
keys. The existing regular expression was not specific enough, and could
result in false positives. This patch fixes the problem by using anchors
to make the regular expression more specific.

Resolves: rhbz#670910

Signed-off-by: Ryan O'Hara <rohara@redhat.com>
Reviewed-by: Lon Hohberger <lhh@redhat.com>