fence_scsi: verify that on/off actions succeed
This patch adds code to verify that both "on" and "off" actions are
successful. This is needed because there seem to be some arrays that
do not report an error even when no registration was created, etc.

These verification steps takes places after the action has been
performed successfully. If an error was encountered while performing
either an "on" or "off" action, fence_scsi will exit and no verification
will take place. Failure to verify that an action was successful for any
device will result is failure.

For the "on" action, do_verify_on will check that the key was
successfully registered with each devices and that a reservation exists
on each device.

For the "off" action, do_verify_off will check that the key was removed
from each device and that a reservation exists on each device.

Resolves: rhbz#644385

Signed-off-by: Ryan O'Hara <rohara@redhat.com>
Reviewed-by: Lon Hohberger <lhh@redhat.com>