fence_scsi: properly log errors for all commands
There are a number of places in fence_scsi that call sg_persist commands
via Perl's qx command. If an error occurs, the script simply dies and
reports the name of the subroutine that the script died in. This can be
improved by replacing the die call with log_error, such that any errors
get properly written to the logfile, if any. This patch also logs the
error code returned from the failed command.

This patch also redirects stderr to /dev/null for all commands executed
via qx.

Signed-off-by: Ryan O'Hara <rohara@redhat.com>
Reviewed-by: Lon Hohberger <lhh@redhat.com>