fence_scsi_check: watchdog script for fence_scsi
This fence_scsi_check watchdog script will check to see is the local
node's key is registered with any devices. If the key is registered with
at least one device, the script will return 0. If none of the devices
have the node's key registered (ie. it has been fenced), the script will
return 2 and node will reboot via watchdog.

Signed-off-by: Ryan O'Hara <rohara@redhat.com>
Reviewed-by: Fabio M. Di Nitto <fdinitto@redhat.com>