Allow fence_scsi to use any valid hexadecimal key
The check for a non-zero key is using perl's implicit string-to-integer
conversion that only works on base 10 digits, which means that any key
starting with at least one digit [1-9] will come through as okay, but
any key starting with [A-F] will evaluate as 0 and fail the test.

An explicit conversion to integer via hex() is the solution.

I hit this in a cluster with a cluster ID of 43316 == 0xA934 and scsi
fencing configured, where starting the cman service would always fail at
the "Unfencing self" step.

Signed-off-by: Jim Ramsay <jim_ramsay@dell.com>
Signed-off-by: Ryan O'Hara <rohara@redhat.com>

Resolves: rhbz#653504