fence_scsi: grep for keys should be case insensitive
When checking that a key is or is not registered for a device, the key
comparison should be case insensitive. Using sg_persist to query for
registered keys will always give lowercase characters for hexadecimal
keys, so registering key 'A' would appear to fail since it would not
match 'a'.

Resolves: rhbz#653504

Signed-off-by: Ryan O'Hara <rohara@redhat.com>