ipmilan: Allow privilege selection
In some cases, it's desirable to limit the rights to an iLO
user in order to mitigate the fact that the login information
is stored as plaintext in the cluster configuration.

Unfortunately, ipmitool and fence_ipmilan do not work in this
configuration.

   # /usr/bin/ipmitool -I lanplus -H '10.1.9.40' \
     -U 'fence' -P 'fencepost' -v chassis power status
   RAKP 2 message indicates an error : unauthorized role requested
   Error: Unable to establish IPMI v2 / RMCP+ session
   Unable to get Chassis Power Status

It turns out that dialing back the permissions on a user in iLO
changes its privilege level, in this case from administrator to
operator.  ipmitool defaults to administrator if nothing is
specified, hence the error.  If the correct privilege level is
given to ipmitool the command works:

   # /usr/bin/ipmitool -I lanplus -H '10.1.9.40' \
     -U 'fence' -L 'OPERATOR' -P 'fencepost' -v chassis power status
   Chassis Power is on

This patch adds the 'privlvl' option to fence_ipmilan, allowing
administrators to change the privilege level in the cluster
configuration.

Signed-off-by: Kristoffer Knigga <kris@knigga.com>
Signed-off-by: Lon Hohberger <lhh@redhat.com>