fence_drac5: Incorrect output of 'list' operation on Drac 5
Previously (due to other bug) fence agents operation 'list' was not
working properly. After fixing that issue, drac5 returned empty line
instead of N/A as an output for 'list'. This happends because fencing
library does not have support for situations when for one fence agents
there are fence devices that supports / not supports 'list' option.

Resolves: rhbz#718194 (regression introduced by previous patch)