fence_kdump: change default timeout back to 60
After more testing, the time between when the fence_kdump agent begins
listening for a message and the time that it receives a message appears
to be approximately 43-45 seconds. The default timeout of 60 seconds
should be sufficient. Users that wish to wait more than 60 seconds can
increase the timeout via the "timeout=<time>" option.

Signed-off-by: Ryan O'Hara <rohara@redhat.com>