fence_scsi: fix simultaneous unfence operations
This patch fixes an issue where multiple nodes attempt to unfence at
the same time and fail when attempting to create a reservation. Each
node checks to see if a reservation exists, and if not, it will attempt
to create one. The problem occurs when multiple nodes see that no
reservation exists and attempt to create one.

This patch checks the return code of do_reserve. If this call fails,
check again to see if a reservation was created by another node, in
which case there is no error.

Resolves: rhbz#738384

Signed-off-by: Ryan O'Hara <rohara@redhat.com>
Reviewed-by: Fabio M. Di Nitto <fdinitto@redhat.com>