fence_scsi: remove unlink of fence_scsi.dev file
This patch removes the call to unlink the fence_scsi.dev file during
unfencing (action=on). Instead of unlinking the file, check to see if
the current device is already in the fence_scsi.dev file before writing
it to the file. Note that since this file exists in /var/run/cluster
directory, it should be removed on reboot.

Resolves: rhbz#741339

Signed-off-by: Ryan O'Hara <rohara@redhat.com>