fence_vmware_soap: Faster fencing, fix crash on VM without valid UUID
Improve speed of fencing by removing requests for attributes that are not needed. This change is significant
when there are hundrens of VM on vSphere server. On the systems with <10 VM improvement is still about 20%.

This patch also fixes situation when there are VM which do not have valid UUID. This can happend when P2V (physical
to virtual machine process) failed.

Patch was proposed by: Rodrigo A B Freire

Resolves: rhbz#836654