fence_drac5: Fix regression on Dell CMC and Dell DRAC5
Standard EOL for agents connecting via ssh is CR/LF.
Some Dell devices represents CR/LF as double-enter what creates a problem in parsing of output.

This patch adds a check for double-enter. This can be detected in function which power on/off machine
because 'get power status' was run before and if we can find a line without any command we know that
there is a double-enter problem.

Resolves: rhbz#904195

Resolves: rhbz#904195