fencing: Improve work with invalid power states
Previously, status for N ports was OFF if all of these nodes where off
and it differs when of these ports were ON. This did not work properly
when fence agents responeded with 'unknown' values. This patches fixes
this issue and if at least one value is non-OFF then status is not-OFF too.