fencing: Some short options were not propagated, so they were removed
In old versions, it was required to have short getopt options (-k, -x, ...) for
each option available. This is not true for some time but these options
were defined also for options were they were not really used and long options were
the only presented.

Main reason for removing this short options from specific options was that some
of them were really overloaded.