[refactor] simplifying function metadata() + fix XML for fence_cisco_ucs
Two if-s where removed as they are no longer needed by fence agents. One of
its impact was that two fence agents has changed their XML metadata. In case
of fence_ovh it was because default value was set even when it should not be.
And in the case of cisco_ucs, default was not propagated properly because
it was empty string.