fencing: Short options can be misinterpreted
If there are several entries in all_opt that have same short getopt. It
is possible that they are handled incorrectly and they are stored under same
key in 'options' variable.

This update fixes this problem for cases when the other of all_opt entries should
not be present at all. E.g. -d is used for vmware_type and snmp_version, and they are
not allowed at same time.