fencing: Add operation 'validate-all'
Previously, user had to run fence agent in non-destructive mode to find out
if parameters are correct or not. With this commit, the action 'validate-all'
(required by OCF) is added and parameters that can be checked without connecting
to real hardware are tested.

As a bonus of this approach, if there are several problems, user will know all
of them at once what is a benefit also for user interfaces.