fence_rhevm: Fill in outlet status in get_list
The outlet status was set to None for all outlets. This results in a
traceback in fencing.py as it tries to execute the upper() function on
the status. This is fine when the account that tries to access the data
is an admin account as the list will be empty. If a UserRole user is
used, the vm list is not empty and the error occurs.

Fill in the status of each vm in the outlet status to prevent the
traceback.