fence_scsi: Status report "off" if any device is off during "on" action
We need status to report "on" if any device is on most of the time so
that stonith or other entities can know if a host is capable of
accessing shared resources, but during an "on" action this is
problematic because even if some devices are off, we will skip turning
them on since we reported the status as on.  This changes the behavior
to report "off" if any device is off, but only during "on" actions.

Resolves: rhbz#1214919