fence_scsi: Offer hard-reboot option for fence_scsi_check script
The existing implementation of fence_scsi_check returns an error if any
device is no longer registered properly, and this error return causes
watchdog to use its custom procedure to reboot the host.  This procedure
is prone to blocking, especially when GFS2 file systems are mounted or
multipath devices are configured to queue indefinitely, so having the
check be able to hard-reboot the host instead of returning a failure
gives a means for avoiding these blockages.