fence_apc: Support for v6.x firmware
There are no changes in interface, so it is enough to accept new version number.

Login process over telnet has changed a bit and autodection of EOL did not work properly. Relevant part
of changed so it work correctly when after empty 'User Name', the device ask for it again.

Resolves: rhbz#1259254