fence_compute: Fix disabling force_down on node when action is on
When the action is on, the goal is to disable force_down for the
nova-compute service on the node.

However, we were only doing that if the nova-compute service was up;
which is impossible if it's forced to be down... So just always disable
force_down, and then, if it's up, do more things.