build: fix possibly undefined macro AC_PYTHON_MODULE
As already mentioned 2 years back by Alan Evangelista[1], fence-agents
project indeed uses a strange build setup, which furthermore makes it
impossible to pass the configure stage successfully, at least with
some combinations of the "autotools" programs' versions (e.g., as old
ones as present in RHEL 6).

Remedy is as simple as originally suggested, which also means m4
directory no longer needs to be reinsured in autogen.sh (though it
still is as a legacy), but special care must be taken with .gitignore
and maintainer-clean-local target of the main Makafile.

[1] https://www.redhat.com/archives/cluster-devel/2014-October/msg00161.html