feature-powerman Add list support and verify power state after changes
Add list support and change log level to error, for error messages users
should be made aware of.

When fencing or unfencing a node, verify the power state of the device
with --query before returning.  Return code from the powerman "on" or
"off" commands can be unreliable for some devices.