feature powerman address review comments pass 2
* Set default for powerman_path in configure

* Remove unnecessary query after powerman --on or --off
fence_action() automatically queries power status after a power_set
function call, because many devices report success when the on/off
actually failed.

* Delay before checking powerman is executable
Use the provided run_delay() function to determine when a delay is
necessary.  Call it before testing that the powerman binary is
executable.  The project's sanity tests depend on this to handle the
missing dependency (powerman), within their testing environment.