fencing: Specify command to be logged by run_command
Currently run_command special cases the logging of commands with ipmitool and removes everything
between the flat -P and -p. Instead of special casing commands with ipmitool allow the caller to
provide the command to log so that the command can do any necessary scrubbing of the message to be
logged.