Explicitly use version 3 of the oVirt API
Version on 4 of the oVirt engine will support two versions of the API:
version 3, which will be backwards compatible with older versions of the
engine, and version 4, which won't be compatible. The default used by
the server will be version 4, but clients can explicitly request a
version of the API using the "Version" header or the "/v3" URL prefix.
The header is the preferred mechanism because it is just ignored by
older versions of the server. This patch modifies the oVirt fence agent
so that the path and version of the API are configurable, and so that
their default values work correctly with version 3 compatibility mode of
the API.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>