build: release.mk: reflect current release publishing practice
In the previous maintenance epoch, fence-agents release files used
to be published at fedorahosted.org[1], but since the sunset of that
hosting[2] (original files now preserved[3] at [4]), only GitHub is
used to host the releases (in autogerated-from-git fashion).

So reflect that change in the "publish" recipe, dropping the previously
used publish-by-scp convenience of fedorahosted.org that won't fly with
GitHub that insists on the (primarily REST) API to get things automated,
if the habit of publishing customized tarballs is ever put in place.

Also, turn "git push --tags" to a bit safer "git push --follow-tags"
for good measure, and make the enumerated announce channels in the
final prompt reflect the current practice.

[1] https://fedorahosted.org/releases/f/e/fence-agents/ (now defunct)
[2] https://lists.fedoraproject.org/archives/list/announce@lists.fedoraproject.org/message/BWSMCGZPPNG3JOCFQ6Z74MIBU7FG3KGB/
[3] http://oss.clusterlabs.org/pipermail/developers/2017-February/000438.html
[4] http://releases.pagure.org/ClusterLabs/fence-agents/

Signed-off-by: Jan Pokorný <jpokorny@redhat.com>