Add a fence_crosslink agent
As suggested by Andrew we came up with this small agent
to cater for a very specific two-node use-case where a crosscable
connects the two nodes and there is no external PDU fencing
available (aka only the integrated BMC is usable).
With this agent we can make the assumption that if the crosscable
is not available the other node has been powered off.

It is meant to be used with the normal BMC/IPMI fencing as a second
level. See the README.md file for more information.

Co-Authored-By: Damien Ciabrini <dciabrin@redhat.com>