fence_apc.py compatibility for Firmware major release 7 workaround #475
My bad, accidentaly removed the last pull request. Since APC/Schneider Electric released major version 7 of the firmware, this is a workaround to keep the script using the "*5s" methods. The implementation seems to still work fine, but I did not perform extensive and detailed tests. Moreover a more robust and safe control on the major version number has to be considered.