fence_docker: Connect to Docker via a UNIX socket
With this change we extend the Docker connection options to connect via
a UNIX socket. To do so we add a new command line argument
--unix-socket which expects the path to the Docker socket.

Also in order for this to make sense we need to set the --ssl flag to
false, otherwise the agent forces an HTTPS connection. By default this
is not feasible as is, thus we add an extra flag, --disable-ssl to do
so.

Closes #581
Closes #582

Signed-off-by: Marjus Cako <marios.cako@hpe.com>